added cast template, custom scrollbar, etc
