+++ components/movieDetails/Recommendation.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function Recommendation({ recommendationList }: any) {
+  return <div>Recommendation</div>
+}
